% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGC.R
\name{is.transportGC}
\alias{is.transportGC}
\title{Check validity of g-computation result object}
\usage{
is.transportGC(transportGCResult)
}
\arguments{
\item{transportGCResult}{Result object from \code{transportGC} function}
}
\value{
A boolean indicating whether all components of \code{transportGC} object have the correct types.
}
\description{
A simple helper function that validates whether the components of the given \code{transportGC} object are of the correct types.
}
