% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGCPreparedModel.R
\name{is.transportGCPreparedModel}
\alias{is.transportGCPreparedModel}
\title{Check validity of prepared model object for g-computation}
\usage{
is.transportGCPreparedModel(preparedModel)
}
\arguments{
\item{preparedModel}{Output object from \code{transportGCPreparedModel} function}
}
\value{
A boolean indicating whether all components of \code{transportGCPreparedModel} object have the correct types.
}
\description{
A simple helper function that validates whether the components of the given \code{transportGCPreparedModel} object are of the correct types.
}
