% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportIP.R
\name{is.transportIP}
\alias{is.transportIP}
\title{Check validity of IOPW result object}
\usage{
is.transportIP(transportIPResult)
}
\arguments{
\item{transportIPResult}{Result object from \code{transportIP} function}
}
\value{
A boolean indicating whether all components of \code{transportIP} object have the correct types.
}
\description{
A simple helper function that validates whether the components of the given \code{transportIP} object are of the correct types.
}
