% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGC.R
\name{plot.transportGC}
\alias{plot.transportGC}
\title{Visually represent results of transportability analysis using g-computation}
\usage{
\method{plot}{transportGC}(x, ...)
}
\arguments{
\item{x}{Result from \code{transportGC} function.}

\item{...}{Further arguments from previous function or to pass to next function}
}
\value{
A \code{ggplot} object showing the estimates and confidence intervals of the MSM coefficients.
}
\description{
This function is a wrapper for \code{modelsummary::modelplot} to plot the coefficient estimates in a transportability analysis using g-computation. Note that the correct variance estimates are used in this function.
}
