% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportGC.R
\name{summary.transportGC}
\alias{summary.transportGC}
\alias{print.summary.transportGC}
\title{Summarize results of a fitted MSM object using g-computation}
\usage{
\method{summary}{transportGC}(object, ...)

\method{print}{summary.transportGC}(x, out = stdout(), ...)
}
\arguments{
\item{object}{Result from \code{transportGC} function}

\item{...}{Further arguments from previous function or to pass to next function}

\item{x}{\code{summary.transportGC} object.}

\item{out}{Output stream.}
}
\value{
The \code{summary.transportGC} function returns a \code{summary.transportGC} object containing the following components:
\itemize{
\item \code{msmSummary}: Summary object of MSM with correct variance estimates.
\item \code{preparedModelSummary}: Summary object of outcome model, provided only for information. No conclusions should be drawn from the outcome model.
\item \code{response}: String indicating response variable name.
\item \code{treatment}: String indicating treatment variable name.
\item \code{treatmentLevels}: Vector of strings indicating levels of treatment variable
}
}
\description{
Returns summary object which contains summary objects for the MSM and the outcome model, as well as information about response and treatment variables. In the MSM summary object, the correct variance estimators are calculated.
}
