% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportIP.R
\name{summary.transportIP}
\alias{summary.transportIP}
\alias{print.summary.transportIP}
\title{Summarize results of a fitted MSM using the IOPW approach}
\usage{
\method{summary}{transportIP}(transportIPResult, covariates = NULL, effectModifiers = NULL)

\method{print}{summary.transportIP}(summaryTransportIP, out = stdout())
}
\arguments{
\item{transportIPResult}{Result from transportIP function}

\item{covariates}{Vector of strings indicating names of covariates in propensity model}

\item{effectModifiers}{Vector of strings indicating names of effect modifiers in participation model}

\item{summaryTransportIP}{Result from transportIP function.}

\item{out}{Output stream.}
}
\value{
The \code{summary.transportIP} function returns a \code{summary.transportIP} object containing the following components:
\itemize{
\item \code{propensitySMD}: Table of unweighted and weighted SMDs of covariates between treatment groups. Only propensity weights are used.
\item \code{participationSMD}: Table of unweighted and weighted SMDs of effect modifiers between study data and target data. Only participation weights are used.
\item \code{msmSummary}: Summary object of model object for MSM. The correct variance estimators are set here for \code{glm}, whereas they are set in \code{transportIP} for \code{survreg} and \code{coxph}.
}
}
\description{
Returns summary object which contains a summary of the fitted MSM as well as pre- and post-weighting standardized mean differences (SMDs).
}
