% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportIP.R
\name{transportIP}
\alias{transportIP}
\title{Transportability analysis using IOPW}
\usage{
transportIP(
  msmFormula,
  propensityScoreModel = NULL,
  participationModel = NULL,
  propensityWeights = NULL,
  participationWeights = NULL,
  treatment = NULL,
  participation = NULL,
  response = NULL,
  family = stats::gaussian,
  data,
  transport = T
)
}
\arguments{
\item{msmFormula}{A formula for the MSM to be fitted, which usually includes the outcome, the treatment and any effect modifiers.}

\item{propensityScoreModel}{Either a formula or a \code{glm} object representing the model for treatment assignment given covariates.}

\item{participationModel}{Either a formula or a \code{glm} object representing the model for study participation given effect modifiers. If a formula and the "data" argument is a list, then add participation as the left-hand side of the formula.}

\item{propensityWeights}{Vector of custom weights balancing covariates between treatments. Providing them will override the formula or model provided by \code{propensityScoreModel}. This vector should have as any entries as the sample size of the study data.}

\item{participationWeights}{Vector of custom weights balancing effect modifiers between study and target populations. Providing them will override the formula or model provided by \code{participationModel}. This vector should have as any entries as the sample size of the study data.}

\item{treatment}{String indicating name of treatment variable. If \code{NULL}, it will be auto-detected from \code{propensityScoreModel} if provided; otherwise it will remain \code{NULL}. Note that when using custom weights, \code{treatment} should be provided so that \code{summary.transportIP} and \code{plot.transportIP} works.}

\item{participation}{String indicating name of participation variable. If \code{NULL}, it will be auto-detected from \code{participationModel} if provided; otherwise it will remain \code{NULL}. Note that when using custom weights, \code{participation} should be provided so that \code{summary.transportIP} and \code{plot.transportIP} works.}

\item{response}{String indicating name of response variable. If \code{NULL}, it will be auto-detected form \code{msmFormula}.}

\item{family}{Either a \code{family} function as used for \code{glm}, or one of \code{c("coxph", "survreg")}.}

\item{data}{Either a single data frame containing merged study and target datasets, or a list containing the study dataset and the target dataset. Note that if participationModel is a glm object, the datasets would have been merged, so provide the merged dataset containing response, treatment, covariates controlled for in the original study, study participation and effect modifiers if this is the case. Make sure to code treatment and participation as 0-1 or TRUE-FALSE, with 1 and TRUE representing treatment group and study data, respectively.}

\item{transport}{A boolean indicating whether a generalizability analysis (false) or a transportability analysis (true) is done.}
}
\value{
A \code{transportIP} object containing the following components:
\itemize{
\item \code{msm}: Raw model fit object for MSM of class \code{glm}, \code{survreg} and \code{coxph}, with the correct variance estimators appropriately replaced
\item \code{propensityScoreModel}: Model of treatment assignment, \code{NULL} if not provided and custom propensity weights are used
\item \code{participationModel}: Model of study participation, \code{NULL} if not provided and custom propensity weights are used
\item \code{propensityWeights}: Propensity weights used
\item \code{participationWeights}: Participation weights used
\item \code{finalWeights}: Weights used to fit MSM
\item \code{customPropensity}: Boolean indicating whether custom propensity weights are used
\item \code{customParticipation}: Boolean indicating whether custom participation weights are used
\item \code{treatment}: String indicating variable name of treatment
\item \code{participation}: String indicating variable name of participation
\item \code{response}: String indicating variable name of response
\item \code{data}: Data provided in \code{data} argument. Either a list containing study data and target data or a data frame containing both.
}
}
\description{
Estimates the coefficients of a marginal structural model (MSM) using IP weighting in a generalizability or transportability analysis. In particular, the estimators should be unbiased for the coefficients in the superpopulation or the target population, respectively.
}
\details{
The function fits models of treatment assignment and study participation in order to calculate the weights used to fit the MSM. For each of these models, if a formula is provided, logistic regression is used by default. If a \code{glm} object is provided, the function extracts the necessary weights from the object. The function does not support other weighting methods, so if they are required, provide custom weights.

The MSM-fitting functions do not provide correct standard errors as-is. \code{sandwich::vcovBS} is used to calculate robust bootstrap variance estimators of the parameter estimators. The function replaces the variance component in \code{summary.glm}, \code{coxph} and \code{survreg} with the robust variance estimators directly. This does not seem to behave well with \code{predict.glm} yet, but prediction is not of primary interest in a generalizability or transportability analysis.
}
