% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportIP.R
\name{plot.transportIP}
\alias{plot.transportIP}
\title{Plot graphs relevant to transportability analysis using IOPW}
\usage{
\method{plot}{transportIP}(transportIPResult, type = "propensityHist")
}
\arguments{
\item{transportIPResult}{Result from transportIP function}

\item{type}{One of \code{"propensityHist", "propensitySMD", "participationHist", "participationSMD", "msm"}. \code{Hist} produces mirrored histograms of estimated probability of treatment between treatment groups (for \code{propensity}), or of estimated probability of participation between study and target data (for \code{participation}). \code{SMD} produces SMD plots of covariates between treatment groups (for \code{propensity}) or effect modifiers between study and target data (for \code{participation}). \code{msm} produces plots showing confidence intervals for the model coefficients, which should have the correct standard errors.}
}
\value{
A \code{ggplot} object which contains the desired plot.
}
\description{
Plot graphs for assessment of covariate balance and results in a IOPW analysis. This function currently supports mirrored histograms, SMD plots and model coefficient plots.
}
